+++ src/platforms/android/__tests__/permission-grant-state.test.ts
@@ -1,9 +1,6 @@
 import { test } from 'vitest';
 import assert from 'node:assert/strict';
-import {
-  androidPriorGrantState,
-  parseAndroidRuntimePermissionGrants,
-} from '../permission-grant-state.ts';
+import { parseAndroidRuntimePermissionGrants } from '../permission-grant-state.ts';
 
 // Captured from `adb shell dumpsys package com.callstack.agentdevicelab` on a Pixel 7 / API 36
 // emulator, trimmed to the sections that decide the answer. The indentation is load-bearing:
@@ -12,9 +9,6 @@ import {
 const DUMPSYS = [
   'Packages:',
   '  Package [com.example.app] (5f3a1c2):',
-  '    userId=10234',
-  '    declared permissions:',
-  '      com.example.app.DYNAMIC_RECEIVER_NOT_EXPORTED_PERMISSION: prot=signature',
   '    install permissions:',
   '      android.permission.MODIFY_AUDIO_SETTINGS: granted=true',
   '      android.permission.RECORD_AUDIO: granted=true',
@@ -36,45 +30,29 @@ const DUMPSYS = [
   '    path: /data/app/~~abc==/com.example.app-def==',
 ].join('\n');
 
-test('runtime grants are read from the requested user only', () => {
-  const user0 = parseAndroidRuntimePermissionGrants(DUMPSYS, 0);
-  assert.equal(androidPriorGrantState(user0, 'android.permission.RECORD_AUDIO'), 'not_granted');
-  assert.equal(androidPriorGrantState(user0, 'android.permission.CAMERA'), 'granted');
-
-  // The same permission, granted for another profile: `pm revoke` acts on the acting user, so
-  // reporting user 10's grant for a user 0 revoke would claim a kill that never happened.
-  const user10 = parseAndroidRuntimePermissionGrants(DUMPSYS, 10);
-  assert.equal(androidPriorGrantState(user10, 'android.permission.RECORD_AUDIO'), 'granted');
-  assert.equal(androidPriorGrantState(user10, 'android.permission.CAMERA'), 'unknown');
+// `undefined` is the answer for anything the dump does not place inside the requested user's
+// runtime block: an install-permission grant (`pm revoke` cannot touch it), another profile's
+// grant, or a permission that user never declared. The caller reports that as `unknown`.
+test.each([
+  [0, 'android.permission.RECORD_AUDIO', 'not_granted', 'acting user overrides the install grant'],
+  [0, 'android.permission.CAMERA', 'granted', 'acting user'],
+  [10, 'android.permission.RECORD_AUDIO', 'granted', 'other profile, only when asked for'],
+  [10, 'android.permission.CAMERA', undefined, 'absent for that user'],
+  [0, 'android.permission.MODIFY_AUDIO_SETTINGS', undefined, 'install-only'],
+  [0, 'android.permission.INTERNET', undefined, 'never mentioned'],
+] as const)('user %s reads %s as %s (%s)', (userId, permission, expected, _why) => {
+  assert.equal(parseAndroidRuntimePermissionGrants(DUMPSYS, userId)?.get(permission), expected);
 });
 
-test('install permissions never answer for a runtime permission', () => {
-  // RECORD_AUDIO appears as `granted=true` in the install-permission section and `granted=false`
-  // in user 0's runtime block. A scan that matched `granted=true` anywhere read the wrong one.
-  const grants = parseAndroidRuntimePermissionGrants(DUMPSYS, 0);
-  assert.equal(androidPriorGrantState(grants, 'android.permission.RECORD_AUDIO'), 'not_granted');
-  assert.equal(
-    androidPriorGrantState(grants, 'android.permission.MODIFY_AUDIO_SETTINGS'),
-    'unknown',
-  );
-  assert.equal(androidPriorGrantState(grants, 'android.permission.INTERNET'), 'unknown');
-});
-
-test('a user with no runtime-permission block is unknown, not empty', () => {
-  assert.equal(parseAndroidRuntimePermissionGrants(DUMPSYS, 11), undefined);
-  assert.equal(
-    androidPriorGrantState(
-      parseAndroidRuntimePermissionGrants(DUMPSYS, 11),
-      'android.permission.RECORD_AUDIO',
-    ),
-    'unknown',
-  );
-});
-
-test('unparseable output is unknown rather than not-granted', () => {
-  for (const output of ['', 'Packages:\n  <none>', 'Error: package not found']) {
-    assert.equal(parseAndroidRuntimePermissionGrants(output, 0), undefined, output);
-  }
+// A missing block is not an empty one. Every input here means "the device did not tell us",
+// which must stay distinguishable from "the app holds nothing".
+test.each([
+  ['no runtime block for that user', DUMPSYS, 11],
+  ['empty output', '', 0],
+  ['no Packages section', 'Activity Resolver Table:\n  User 0:\n    runtime permissions:', 0],
+  ['a package section without the user', 'Packages:\n  Package [com.example.app] (abc):', 0],
+] as const)('%s reads as unknown', (_label, output, userId) => {
+  assert.equal(parseAndroidRuntimePermissionGrants(output, userId), undefined);
 });
 
 test('an empty runtime block answers not_granted for everything it could have listed', () => {
@@ -94,23 +72,6 @@ test('an empty runtime block answers not_granted for everything it could have li
   assert.equal(grants?.size, 0);
 });
 
-test('a dump without a Packages section is unknown', () => {
-  // Every `dumpsys package <pkg>` carries one (it sits after the resolver tables and Key Set
-  // Manager). Its absence means the output is not the dump we can read, so the honest answer
-  // is unknown — never the not_granted that would claim the app was left alone.
-  const grants = parseAndroidRuntimePermissionGrants(
-    [
-      'Activity Resolver Table:',
-      '  Non-Data Actions:',
-      '    User 0: installed=true',
-      '      runtime permissions:',
-      '        android.permission.RECORD_AUDIO: granted=true, flags=[ USER_SET]',
-    ].join('\n'),
-    0,
-  );
-  assert.equal(grants, undefined);
-});
-
 test('sections after Packages: cannot reopen the scan', () => {
   // `Queries:` repeats `User 0:` with no grants, and `Shared users:` repeats runtime grant lines
   // for the shared uid. Merging either would flip RECORD_AUDIO to granted for a user whose own
@@ -127,6 +88,6 @@ test('sections after Packages: cannot reopen the scan', () => {
     0,
   );
 
-  assert.equal(androidPriorGrantState(grants, 'android.permission.RECORD_AUDIO'), 'not_granted');
-  assert.equal(androidPriorGrantState(grants, 'android.permission.CAMERA'), 'granted');
+  assert.equal(grants?.get('android.permission.RECORD_AUDIO'), 'not_granted');
+  assert.equal(grants?.get('android.permission.CAMERA'), 'granted');
 });
